s/unsigned/unsigned long/ (amd64 fixes, damir@altlinux)
git-svn-id: file:///.data/svn_convert/svn@5798 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/sim/plugins/_core/cfgdlg.cpp
@@ -309,7 +309,7 @@ ConfigureDialog::ConfigureDialog()
 ConfigureDialog::~ConfigureDialog()
 {
     lstBox->clear();
-    for (unsigned n = 0;; n++){
+    for (unsigned long n = 0;; n++){
         Event e(EventPluginGetInfo, (void*)n);
         pluginInfo *info = (pluginInfo*)e.process();
         if (info == NULL) break;
@@ -353,7 +353,7 @@ void ConfigureDialog::fill(unsigned id)
         }
     }
 
-    unsigned n;
+    unsigned long n;
     parentItem = NULL;
     list<unsigned> st;
     for (n = 0; n < getContacts()->nClients(); n++){
